requirements.txt
flask==3.0.3
flask-cors==4.0.0
gunicorn==22.0.0
supabase==2.4.6
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
requests==2.31.0
schedule==1.2.1
click==8.1.7
tabulate==0.9.0
python-json-logger==2.0.7

